Pull in github pages actions and reformat
I wanted to set up github pages to make my example code comprehensible. In the process I found that some classes had not been renamespaced when moved to a new folder. Naturally I was forced to go on an adventure to set up an editorconfig to prevent this and it found many other problems too, which are now fixed